Al Rouf Tasks/TASK 3 - RAG Knowledge Base/requirements.txt
faiss-cpu
sentence-transformers
langdetect
flask
openai
numpy
